feat: hocr structured ocr result, migrate db schema
